fix(ci): clang-format 18, clang-tidy narrowing/init, Windows MSVC runtime
Resolves all three failing CI checks on the Phase 2 PR.

- clang-format: reformat to clang-format 18.1.8 (the version ubuntu-latest
  now ships); the bump reflows several files, including pre-existing ones.
- clang-tidy: fix 5 warnings-as-errors findings:
  - Engine.cpp / UISystem.cpp: explicit static_cast<float> on int->float
    narrowing (bugprone-narrowing-conversions)
  - PlayerController.cpp: initialize m_dashCharges in the constructor member
    initializer list (cppcoreguidelines-prefer-member-initializer)
- Windows: set USE_STATIC_MSVC_RUNTIME_LIBRARY OFF before Jolt's
  add_subdirectory. Jolt defaulted to the static /MT runtime while every
  other target uses the dynamic /MD runtime, causing LNK2038 on every
  Jolt-linked executable (DoomScroller.exe, level_convert.exe). This has
  failed on main since the first Jolt-linked exe landed.

+++ engine/src/Engine.cpp
@@ -1177,7 +1177,7 @@ void Engine::renderHUD(rhi::IRHICommandList* cmd) {
     m_ui.begin(m_windowWidth, m_windowHeight);
 
     const auto& health   = m_world.get<HealthComponent>(m_playerEntity);
-    float healthFrac     = health.max > 0 ? static_cast<float>(health.current) / health.max : 0.f;
+    float healthFrac     = health.max > 0 ? static_cast<float>(health.current) / static_cast<float>(health.max) : 0.f;
     healthFrac           = std::clamp(healthFrac, 0.f, 1.f);
     const bool lowHealth = healthFrac <= 0.3f;
